requirements.txt
functions-framework==3.3.0
python-dotenv==1.0.0
requests==2.31.0
mysql-connector-python==8.1.0
sqlalchemy==2.0.21